Rename AppDomainAdapter to AssemblyProvider to match purpose
