--- src/UnityContainerAttributeRegistration/Adapter/AppDomainAdapter.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
-
-
-namespace UnityContainerAttributeRegistration.Adapter
-{
-    /// <inheritdoc cref="IAppDomainAdapter" />
-    internal class AppDomainAdapter : IAppDomainAdapter
-    {
-        /// <inheritdoc cref="IAppDomainAdapter.GetAssemblies" />
-        public IList<Assembly> GetAssemblies()
-        {
-            return AppDomain.CurrentDomain.GetAssemblies();
-        }
-    }
-}
