+++ src/UnityContainerAttributeRegistration/Provider/IAssemblyProvider.cs
@@ -2,12 +2,12 @@
 using System.Reflection;
 
 
-namespace UnityContainerAttributeRegistration.Adapter
+namespace UnityContainerAttributeRegistration.Provider
 {
     /// <summary>
     ///     Wrapper to provide <see cref="IList{T}" /> instead of using <see cref="System.AppDomain" />.
     /// </summary>
-    public interface IAppDomainAdapter
+    public interface IAssemblyProvider
     {
         /// <summary>
         ///     Provide <see cref="Assembly" />s of an AppDomain.
